+++ ledger/api/managers/core_manager.py
@@ -3,6 +3,7 @@ from dataclasses import asdict, dataclass, field
 from datetime import datetime
 
 from django.db.models import Q
+from django.utils import timezone
 
 from ledger import app_settings
 from ledger.hooks import get_extension_logger
@@ -45,7 +46,7 @@ class LedgerDate:
     year: int
     month: int
     monthly: bool = field(init=False)
-    current_date: datetime = field(default_factory=datetime.now, init=False)
+    current_date: datetime = None
     range_data: int = field(init=False)
     day_checks: list = field(init=False)
 
@@ -54,6 +55,7 @@ class LedgerDate:
         return num_days
 
     def __post_init__(self):
+        self.current_date = timezone.now()
         self.monthly = self.month == 0
         self.range_data = 12 if self.monthly else self.calculate_days()
         self.day_checks = list(range(1, self.range_data + 1))
@@ -79,9 +81,6 @@ class LedgerFilterCore:
     filter_first_party: Q = field(init=False)
     filter_second_party: Q = field(init=False)
     filter_partys: Q = field(init=False)
-    filter_bounty: Q = field(init=False)
-    filter_ess: Q = field(init=False)
-    filter_mining: Q = field(init=False)
     filter_total: Q = field(init=False)
 
     def __post_init__(self):
@@ -89,6 +88,14 @@ class LedgerFilterCore:
         self.filter_second_party = Q(second_party_id__in=self.char_id)
         self.filter_partys = self.filter_first_party | self.filter_second_party
         self.filter_total = self.filter_partys & Q(amount__gt=0)
+
+
+@dataclass
+class LedgerFilterPvE(LedgerFilterCore):
+
+    def __init__(self, char_id):
+        super().__init__(char_id)
+
         self.filter_bounty = self.filter_second_party & Q(ref_type="bounty_prizes")
         self.filter_ess = self.filter_second_party & Q(ref_type="ess_escrow_transfer")
         self.filter_mining = (
@@ -97,13 +104,33 @@ class LedgerFilterCore:
             else Q(character__character__character_id__in=self.char_id)
         )
 
+        self.filter_all_pve = self.filter_bounty | self.filter_ess | self.filter_mining
 
+    # TODO add mining filter with aggregate
+    def get_all_pve_filters(self):
+        return {
+            "bounty": self.filter_bounty,
+            # "ess": self.filter_ess,
+        }
+
+
+# pylint: disable=too-many-instance-attributes
 @dataclass
-class LedgerFilterCost(LedgerFilterCore):
+class LedgerFilterCost(LedgerFilterPvE):
     """LedgerFilter class to store the filter data."""
 
     def __init__(self, char_id):
         super().__init__(char_id)
+        self.filter_contract_cost = self.filter_partys & Q(
+            ref_type__in=[
+                "contract_price_payment_corp",
+                "contract_reward",
+                "contract_price",
+                "contract_collateral",
+                "contract_reward_deposited",
+            ],
+            amount__lt=0,
+        )
         self.filter_market_cost = self.filter_partys & Q(
             ref_type__in=[
                 "market_escrow",
@@ -113,10 +140,79 @@ class LedgerFilterCost(LedgerFilterCore):
             ],
             amount__lt=0,
         )
-        self.filter_production = self.filter_partys & Q(
-            ref_type__in=["industry_job_tax", "manufacturing"]
+        self.filter_assets_cost = self.filter_partys & Q(
+            ref_type__in=[
+                "asset_safety_recovery_tax",
+            ],
+            amount__lt=0,
         )
+        self.filter_traveling_cost = self.filter_partys & Q(
+            ref_type__in=[
+                "structure_gate_jump",
+                "jump_clone_activation_fee",
+            ],
+            amount__lt=0,
+        )
+        self.filter_production_cost = self.filter_partys & Q(
+            ref_type__in=[
+                "industry_job_tax",
+                "manufacturing",
+                "researching_time_productivity",
+                "researching_material_productivity",
+                "copying",
+                "reprocessing_tax",
+                "reaction",
+            ],
+            amount__lt=0,
+        )
+
+        self.filter_skill_cost = self.filter_partys & Q(
+            ref_type__in=[
+                "skill_purchase",
+            ],
+            amount__lt=0,
+        )
+
+        self.filter_insurance_cost = self.filter_partys & Q(
+            ref_type__in=[
+                "insurance",
+            ],
+            amount__lt=0,
+        )
+
+        self.filter_planetary_cost = self.filter_partys & Q(
+            ref_type__in=[
+                "planetary_import_tax",
+                "planetary_export_tax",
+                "planetary_construction",
+            ],
+            amount__lt=0,
+        )
+
         self.filter_costs = self.filter_partys & Q(amount__lt=0)
+
+        self.filter_all_costs = (
+            self.filter_contract_cost
+            | self.filter_market_cost
+            | self.filter_assets_cost
+            | self.filter_traveling_cost
+            | self.filter_production_cost
+            | self.filter_skill_cost
+            | self.filter_insurance_cost
+            | self.filter_planetary_cost
+        )
+
+    def get_all_costs_filters(self):
+        return {
+            "market_cost": self.filter_market_cost,
+            "production_cost": self.filter_production_cost,
+            "contract_cost": self.filter_contract_cost,
+            "traveling_cost": self.filter_traveling_cost,
+            "asset_cost": self.filter_assets_cost,
+            "skill_cost": self.filter_skill_cost,
+            "insurance_cost": self.filter_insurance_cost,
+            "planetary_cost": self.filter_planetary_cost,
+        }
 
 
 @dataclass
@@ -134,9 +230,29 @@ class LedgerFilterTrading(LedgerFilterCost):
             ],
             amount__gt=0,
         )
+
         self.filter_donation = self.filter_partys & Q(
             ref_type="player_donation", amount__gt=0
         )
+
+        self.filter_insurance = self.filter_second_party & Q(
+            ref_type__in=[
+                "insurance",
+            ],
+            amount__gt=0,
+        )
+
+        self.filter_all_misc = (
+            self.filter_market | self.filter_contract | self.filter_insurance
+        )
+
+    def get_all_misc_filters(self, chars_list):
+        return {
+            "transaction": self.filter_market,
+            "contract": self.filter_contract,
+            "donation": self.filter_donation & ~Q(first_party_id__in=chars_list),
+            "insurance": self.filter_insurance,
+        }
 
 
 @dataclass
@@ -153,6 +269,18 @@ class LedgerFilterMission(LedgerFilterTrading):
             amount__gt=0,
         )
 
+        self.loyality_points_cost = self.filter_partys & Q(
+            ref_type="lp_store", amount__lt=0
+        )
+
+        self.filter_all_missions = self.filter_mission | self.loyality_points_cost
+
+    def get_all_mission_filters(self):
+        return {
+            "mission": self.filter_mission,
+            "loyality_point_cost": self.loyality_points_cost,
+        }
+
 
 @dataclass
 class LedgerFilter(LedgerFilterMission):
@@ -161,3 +289,11 @@ class LedgerFilter(LedgerFilterMission):
     def __init__(self, char_id):
         super().__init__(char_id)
         self.char_id = char_id
+
+    def get_all_filters(self, chars_list):
+        filters = {}
+        filters.update(self.get_all_pve_filters())
+        filters.update(self.get_all_costs_filters())
+        filters.update(self.get_all_misc_filters(chars_list))
+        filters.update(self.get_all_mission_filters())
+        return filters